+++ src/config.js
@@ -2,7 +2,7 @@ module.exports = {
     /** valid vscode version */
     minCodeVersion: '1.60.0',
     /** inquirer choices*/
-    promptChoices: ['js', 'vue2', 'vue3', 'ts', 'vue2-ts', 'vue3-ts'],
+    promptChoices: ['js', 'vue2', 'vue3', 'ts', 'vue2-ts', 'vue3-ts', 'svelte3'],
     /** type -> vscode extensions */
     typeExtMapper: {
         default: ['eslint'],
@@ -11,21 +11,25 @@ module.exports = {
         vue3: ['volar'],
         'vue2-ts': ['vetur'],
         'vue3-ts': ['volar'],
+        svelte3: ['svelte'],
     },
     // vscode extensions
     codeExt: {
         eslint: { name: 'dbaeumer.vscode-eslint', validVersion: '2.2.0' },
         vetur: { name: 'octref.vetur', validVersion: '0.34.1' },
         volar: { name: 'johnsoncodehk.volar', validVersion: '0.33.2' },
+        svelte: { name: 'svelte.svelte-vscode', validVersion: '105.10.0' },
     },
     // ---npm packages
     npmPkgs: {
-        default: {
+        eslint: {
             eslint: '8.10.0',
             'eslint-config-prettier': '8.5.0',
             'eslint-plugin-html': '6.2.0',
-            'eslint-plugin-prettier': '4.0.0',
+        },
+        prettier: {
             prettier: '2.5.1',
+            'eslint-plugin-prettier': '4.0.0',
         },
         js: {},
         ts: {
@@ -49,6 +53,9 @@ module.exports = {
             '@typescript-eslint/eslint-plugin': '5.14.0',
             '@typescript-eslint/parser': '5.14.0',
             '@vue/eslint-config-typescript': '8.0.0',
+        },
+        svelte3: {
+            'eslint-plugin-svelte3': '3.4.1',
         },
     },
     // platform - settings.json path