+++ test/updateSettings.js
@@ -1,5 +1,5 @@
 try {
-    require('../src/updateSettings.js')()
+    require('../src/updateSettings.js')({ type: 'svelte3' })
 } catch (err) {
     console.error(err)
 }